Added the test for the amout of pokemons and splite the main file into modules
