+++ src/modules/amountOfComments.js
@@ -1,0 +1,6 @@
+function amountOfComments(totalComments, pokemonID) {
+  const spanElement = document.querySelector(`.amount-${pokemonID}`);
+  spanElement.textContent = totalComments;
+}
+
+export default amountOfComments;
