+++ src/styles.css
@@ -206,7 +206,7 @@ footer a {
   border-radius: 10px;
   box-shadow: 0 2px 5px rgba(0, 0, 0, 0.3);
   height: 80%;
-  width: 65%;
+  width: 50%;
 }
 
 .popup-header {
@@ -215,19 +215,19 @@ footer a {
 }
 
 .popup-body img {
-  width: fit-content;
-  height: 300px;
+  height: 15rem;
   align-self: center;
 }
 
 .popup-body {
   display: flex;
   flex-direction: column;
+
 }
 
 .pokemon-info {
   display: grid;
-  grid-template-columns: repeat(2, 1fr);
+  grid-template-columns: repeat(3, 1fr);
 }
 
 .popup h2 {
@@ -241,7 +241,8 @@ footer a {
   margin-bottom: 10px;
 }
 
-.popup button {
+.popup button,
+button[type="submit"] {
   background-color: #4caf50;
   color: #fff;
   padding: 10px 20px;
@@ -250,12 +251,9 @@ footer a {
   cursor: pointer;
 }
 
-.comment-section form button {
-  background-color: #4caf50;
-  color: white;
-  border: none;
-  padding: 10px;
-  cursor: pointer;
+.form-section,
+.comme {
+  width: 100%;
 }
 
 .popup button:hover {
@@ -263,17 +261,19 @@ footer a {
 }
 
 .comment-section {
-  margin-top: 20px;
+  height: 8rem;
+  overflow-y: auto;
+  padding: 1rem 0;
 }
 
 .comment-section h3 {
-  margin-top: 0;
+  padding-bottom: 0.5rem;
 }
 
 .comment-section form {
-  margin-top: 10px;
+  margin-top: 0.5rem;
   display: flex;
-  gap: 15px;
+  gap: 0.5rem;
 }
 
 input {
@@ -282,9 +282,9 @@ input {
 
 input,
 textarea {
-  width: 300px;
+  width: 20rem;
   border: 2px solid #070707;
-  height: 30px;
+  height: 2rem;
 }
 
 .comments {
@@ -322,8 +322,5 @@ textarea {
   align-items: center;
   justify-content: center;
   text-align: center;
-}
-
-.comme div {
-  display: flex;
+  padding: 0.8rem 0;
 }
